requirements.txt
fastapi
mysql-connector-python
uvicorn[standard]

-e .